rename regions_parser
